Remove seperate UserPermission-module in app
