+++ demo/api/src/auth/auth.module.ts
@@ -1,7 +1,21 @@
-import { createAuthResolver, createCometAuthGuard, createUserPermissionsStaticAuthedUserStrategy, CurrentUser } from "@comet/cms-api";
+import { createAuthResolver, createCometAuthGuard, createUserPermissionsStaticAuthedUserStrategy, CurrentUser, User } from "@comet/cms-api";
 import { Module } from "@nestjs/common";
 import { APP_GUARD } from "@nestjs/core";
-import { staticUsers } from "@src/user-permission/config.service";
+
+export const staticUsers: User[] = [
+    {
+        id: "1",
+        name: "Admin",
+        email: "demo@comet-dxp.com",
+        language: "en",
+    },
+    {
+        id: "2",
+        name: "Non-Admin",
+        email: "test@test.com",
+        language: "en",
+    },
+];
 
 @Module({
     providers: [
