--- demo/api/src/user-permission/user-permission.module.ts
@@ -1,10 +1,0 @@
-import { Module } from "@nestjs/common";
-
-import { UserPermissionConfigService } from "./config.service";
-
-@Module({
-    imports: [],
-    providers: [UserPermissionConfigService],
-    exports: [UserPermissionConfigService],
-})
-export class UserPermissionModule {}
