+++ packages/api/cms-api/src/user-permissions/user-permissions.types.ts
@@ -26,8 +26,8 @@ export const USERPERMISSIONS_CONFIG_SERVICE = "userpermissions-config-service";
 export const ACCESS_CONTROL_SERVICE = "access-control-service";
 
 export interface UserPermissionConfigInterface {
-    getUser: (id: string) => Promise<User>;
-    findUsers: (args: FindUsersArgs) => Promise<[User[], number]>;
+    getUser: (id: string) => Promise<User> | User;
+    findUsers: (args: FindUsersArgs) => Promise<[User[], number]> | [User[], number];
     getAvailablePermissions?: () => Promise<Permission> | Permission;
     getAvailableContentScopes?: () => Promise<ContentScopeValues> | ContentScopeValues;
     getPermissionsForUser?: (user: User) => Promise<UserPermissions> | UserPermissions;
